initial commit
This is a small service that uses a path as a way of mapping input query args
to an output URL format and set of variable reformatters. This is useful for
an advanced form of what Apache HTTPd URL rewriting does, but with logic
for the parameter value reformatting (eg. changing from epoch to iso-8601 date formats).